+++ acceptance/aws/database_test.go
@@ -95,4 +95,25 @@ func TestDatabase(t *testing.T) {
 		}).ExpectCalls("DescribeDBInstances").Run(t)
 	})
 
+	t.Run("start", func(t *testing.T) {
+		Template("start database id=db-1234").
+			Mock(&rdsMock{
+				StartDBInstanceFunc: func(param0 *rds.StartDBInstanceInput) (*rds.StartDBInstanceOutput, error) {
+					return nil, nil
+				},
+			}).ExpectInput("StartDBInstance", &rds.StartDBInstanceInput{
+			DBInstanceIdentifier: String("db-1234"),
+		}).ExpectCalls("StartDBInstance").Run(t)
+	})
+
+	t.Run("stop", func(t *testing.T) {
+		Template("stop database id=db-1234").
+			Mock(&rdsMock{
+				StopDBInstanceFunc: func(param0 *rds.StopDBInstanceInput) (*rds.StopDBInstanceOutput, error) {
+					return nil, nil
+				},
+			}).ExpectInput("StopDBInstance", &rds.StopDBInstanceInput{
+			DBInstanceIdentifier: String("db-1234"),
+		}).ExpectCalls("StopDBInstance").Run(t)
+	})
 }