+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ### Features
 
+### AWS Services
+
 - Create an image from an existing instance. See `awless create image -h`
     * `awless create image instance=@my-instance-name name=redis-image  description='redis prod image'`
     * `awless create image instance=i-0ee436a45561c04df name=redis-image reboot=true`
     * List your images with `awless ls images --sort created`
     * Delete images with an `awless revert ...` or with `awless delete image id=@redis-image`
+- [#169](https://github.com/wallix/awless/issues/169): Start/Stop a RDS database:
+    * `awless start database id=my-db-id`
+    * `awless stop database id=@my-db-name`
 
 ## v0.1.8 [2017-11-29]